requirements/base.txt
Django>=1.8
djangorestframework
django-filter
markdown
